chore: rename drawer context in 989aa51 & add comment
